refactor(renomeando o arquivo user em /routes): renomeando de users para user
